+++ pkg/controller/reconciler/reconciler_support.go
@@ -4,10 +4,13 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	"github.com/hashicorp/go-multierror"
+	"github.com/opendatahub-io/opendatahub-operator/v2/pkg/controller/types"
+	"github.com/opendatahub-io/opendatahub-operator/v2/pkg/metadata/annotations"
+	"sigs.k8s.io/controller-runtime/pkg/source"
 	"strings"
 
 	"golang.org/x/exp/slices"
-	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apimachinery/pkg/runtime/schema"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/builder"
@@ -15,50 +18,98 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/handler"
 	"sigs.k8s.io/controller-runtime/pkg/predicate"
 
-	"github.com/opendatahub-io/opendatahub-operator/v2/apis/components"
 	"github.com/opendatahub-io/opendatahub-operator/v2/apis/services"
 	"github.com/opendatahub-io/opendatahub-operator/v2/pkg/controller/actions"
 	"github.com/opendatahub-io/opendatahub-operator/v2/pkg/controller/handlers"
 	"github.com/opendatahub-io/opendatahub-operator/v2/pkg/controller/predicates/component"
 	"github.com/opendatahub-io/opendatahub-operator/v2/pkg/controller/predicates/generation"
 	"github.com/opendatahub-io/opendatahub-operator/v2/pkg/metadata/labels"
+	"github.com/opendatahub-io/opendatahub-operator/v2/pkg/resources"
+)
+
+var (
+	// defaultPredicate is the default set of predicates associated to
+	// resources when there is no specific predicate configured via the
+	// builder.
+	//
+	// It would trigger a reconciliation if either the generation or
+	// metadata (labels, annotations) have changed.
+	defaultPredicate = predicate.Or(
+		generation.New(),
+		predicate.LabelChangedPredicate{},
+		predicate.AnnotationChangedPredicate{},
+	)
 )
 
 type forInput struct {
 	object  client.Object
 	options []builder.ForOption
+	gvk     schema.GroupVersionKind
 }
+
+type DynamicPredicate func(context.Context, *types.ReconciliationRequest) bool
 
 type watchInput struct {
 	object       client.Object
 	eventHandler handler.EventHandler
-	options      []builder.WatchesOption
+	predicates   []predicate.Predicate
+	owned        bool
+	dynamic      bool
+	dynamicPred  []DynamicPredicate
 }
-type ownInput struct {
-	object  client.Object
-	options []builder.OwnsOption
+
+type WatchOpts func(*watchInput)
+
+func WithPredicates(values ...predicate.Predicate) WatchOpts {
+	return func(a *watchInput) {
+		a.predicates = append(a.predicates, values...)
+	}
+}
+func WithEventHandler(value handler.EventHandler) WatchOpts {
+	return func(a *watchInput) {
+		a.eventHandler = value
+	}
+}
+
+func WithEventMapper(value handler.MapFunc) WatchOpts {
+	return func(a *watchInput) {
+		a.eventHandler = handler.EnqueueRequestsFromMapFunc(value)
+	}
+}
+
+func Dynamic(predicates ...DynamicPredicate) WatchOpts {
+	return func(a *watchInput) {
+		a.dynamic = true
+		a.dynamicPred = slices.Clone(predicates)
+	}
 }
 
 type ReconcilerBuilder struct {
 	mgr          ctrl.Manager
 	input        forInput
 	watches      []watchInput
-	owns         []ownInput
 	predicates   []predicate.Predicate
 	ownerName    string
 	instanceName string
 	actions      []actions.Fn
 	finalizers   []actions.Fn
+	errors       error
 }
 
 func ReconcilerFor(mgr ctrl.Manager, ownerName string, object client.Object, opts ...builder.ForOption) *ReconcilerBuilder {
 	crb := ReconcilerBuilder{
-		mgr:       mgr,
-		ownerName: ownerName,
-		input: forInput{
-			object:  object,
-			options: slices.Clone(opts),
-		},
+		mgr: mgr,
+	}
+
+	gvk, err := mgr.GetClient().GroupVersionKindFor(object)
+	if err != nil {
+		crb.errors = multierror.Append(crb.errors, fmt.Errorf("unable to determine GVK: %w", err))
+	}
+
+	crb.input = forInput{
+		object:  object,
+		options: slices.Clone(opts),
+		gvk:     gvk,
 	}
 
 	return &crb
@@ -79,60 +130,62 @@ func (b *ReconcilerBuilder) WithFinalizer(value actions.Fn) *ReconcilerBuilder {
 	return b
 }
 
-func (b *ReconcilerBuilder) Watches(object client.Object, opts ...builder.WatchesOption) *ReconcilerBuilder {
-	var instanceLabel string
-	if _, ok := b.input.object.(components.ComponentObject); ok {
-		instanceLabel = labels.ComponentPartOf
-	} else if _, ok := b.input.object.(components.ComponentObject); ok {
-		instanceLabel = labels.ServicePartOf
+func (b *ReconcilerBuilder) Watches(object client.Object, opts ...WatchOpts) *ReconcilerBuilder {
+	in := watchInput{}
+	in.object = object
+	in.owned = false
+
+	for _, opt := range opts {
+		opt(&in)
 	}
-	b.watches = append(b.watches, watchInput{
-		object:       object,
-		eventHandler: handlers.LabelToName(instanceLabel),
-		options:      slices.Clone(opts),
-	})
+
+	if in.eventHandler == nil {
+		// use the platform.opendatahub.io/instance.name label to find out
+		// the owner
+		in.eventHandler = handlers.AnnotationToName(annotations.InstanceName)
+	}
+
+	if len(in.predicates) == 0 {
+		in.predicates = append(in.predicates, predicate.And(
+			DefaultPredicate,
+			// use the components.opendatahub.io/part-of label to filter
+			// events not related to the owner type
+			component.ForLabel(labels.ComponentPartOf, strings.ToLower(b.input.gvk.Kind)),
+		))
+	}
+
+	b.watches = append(b.watches, in)
 
 	return b
 }
 
-func (b *ReconcilerBuilder) WatchesH(object client.Object, eventHandler handler.EventHandler, opts ...builder.WatchesOption) *ReconcilerBuilder {
-	b.watches = append(b.watches, watchInput{
-		object:       object,
-		eventHandler: eventHandler,
-		options:      slices.Clone(opts),
-	})
-
-	return b
+func (b *ReconcilerBuilder) WatchesGVK(gvk schema.GroupVersionKind, opts ...WatchOpts) *ReconcilerBuilder {
+	return b.Watches(resources.GvkToUnstructured(gvk), opts...)
 }
 
-func (b *ReconcilerBuilder) WatchesGVK(gvk schema.GroupVersionKind, eventHandler handler.EventHandler, opts ...builder.WatchesOption) *ReconcilerBuilder {
-	u := unstructured.Unstructured{}
-	u.SetGroupVersionKind(gvk)
+func (b *ReconcilerBuilder) Owns(object client.Object, opts ...WatchOpts) *ReconcilerBuilder {
+	in := watchInput{}
+	in.object = object
+	in.owned = true
 
-	b.watches = append(b.watches, watchInput{
-		object:       &u,
-		eventHandler: eventHandler,
-		options:      slices.Clone(opts),
-	})
+	for _, opt := range opts {
+		opt(&in)
+	}
 
-	return b
-}
+	if in.eventHandler == nil {
+		in.eventHandler = handler.EnqueueRequestForOwner(
+			b.mgr.GetScheme(),
+			b.mgr.GetRESTMapper(),
+			b.input.object,
+			handler.OnlyControllerOwner(),
+		)
+	}
 
-func (b *ReconcilerBuilder) WatchesM(object client.Object, fn handler.MapFunc, opts ...builder.WatchesOption) *ReconcilerBuilder {
-	b.watches = append(b.watches, watchInput{
-		object:       object,
-		eventHandler: handler.EnqueueRequestsFromMapFunc(fn),
-		options:      slices.Clone(opts),
-	})
+	if len(in.predicates) == 0 {
+		in.predicates = append(in.predicates, DefaultPredicate)
+	}
 
-	return b
-}
-
-func (b *ReconcilerBuilder) Owns(object client.Object, opts ...builder.OwnsOption) *ReconcilerBuilder {
-	b.owns = append(b.owns, ownInput{
-		object:  object,
-		options: slices.Clone(opts),
-	})
+	b.watches = append(b.watches, in)
 
 	return b
 }
@@ -142,27 +195,21 @@ func (b *ReconcilerBuilder) WithEventFilter(p predicate.Predicate) *ReconcilerBu
 	return b
 }
 
-func (b *ReconcilerBuilder) BuildComponent(ctx context.Context) (*ComponentReconciler, error) {
+func (b *ReconcilerBuilder) OwnsGVK(gvk schema.GroupVersionKind, opts ...WatchOpts) *ReconcilerBuilder {
+	return b.Owns(resources.GvkToUnstructured(gvk), opts...)
+}
+
+func (b *ReconcilerBuilder) BuildComponent(_ context.Context) (*ComponentReconciler, error) {
+	if b.errors != nil {
+		return nil, b.errors
+	}
+
 	name := b.instanceName
 	if name == "" {
-		kinds, _, err := b.mgr.GetScheme().ObjectKinds(b.input.object)
-		if err != nil {
-			return nil, err
-		}
-		if len(kinds) != 1 {
-			return nil, fmt.Errorf("expected exactly one kind of object, got %d", len(kinds))
-		}
-
-		name = kinds[0].Kind
-		name = strings.ToLower(name)
+		name = strings.ToLower(b.input.gvk.Kind)
 	}
 
-	obj, ok := b.input.object.(components.ComponentObject)
-	if !ok {
-		return nil, errors.New("invalid type for object")
-	}
-
-	r, err := NewComponentReconciler(ctx, b.mgr, name, obj)
+	r, err := NewComponentReconciler(b.mgr, name, b.input.object)
 	if err != nil {
 		return nil, fmt.Errorf("failed to create reconciler for component %s: %w", name, err)
 	}
@@ -183,34 +230,28 @@ func (b *ReconcilerBuilder) BuildComponent(ctx context.Context) (*ComponentRecon
 	c = c.For(b.input.object, forOpts...)
 
 	for i := range b.watches {
-		watchOpts := b.watches[i].options
-		if len(watchOpts) == 0 {
-			watchOpts = append(watchOpts, builder.WithPredicates(predicate.And(
-				generation.New(),
-				component.ForLabel(labels.ComponentPartOf, b.ownerName),
-			)))
+		if b.watches[i].owned {
+			kinds, _, err := b.mgr.GetScheme().ObjectKinds(b.watches[i].object)
+			if err != nil {
+				return nil, err
+			}
+
+			for i := range kinds {
+				r.AddOwnedType(kinds[i])
+			}
 		}
 
-		c = c.Watches(b.watches[i].object, b.watches[i].eventHandler, watchOpts...)
-	}
-
-	for i := range b.owns {
-		ownOpts := b.owns[i].options
-		if len(ownOpts) == 0 {
-			ownOpts = append(ownOpts, builder.WithPredicates(predicate.And(
-				generation.New(),
-			)))
+		// if the watch is dynamic, then the watcher will be registered
+		// at later stage
+		if b.watches[i].dynamic {
+			continue
 		}
 
-		c = c.Owns(b.owns[i].object, ownOpts...)
-		kinds, _, err := b.mgr.GetScheme().ObjectKinds(b.owns[i].object)
-		if err != nil {
-			return nil, err
-		}
-
-		for i := range kinds {
-			r.AddOwnedType(kinds[i])
-		}
+		c = c.Watches(
+			b.watches[i].object,
+			b.watches[i].eventHandler,
+			builder.WithPredicates(b.watches[i].predicates...),
+		)
 	}
 
 	for i := range b.predicates {
@@ -224,7 +265,22 @@ func (b *ReconcilerBuilder) BuildComponent(ctx context.Context) (*ComponentRecon
 		r.AddFinalizer(b.finalizers[i])
 	}
 
-	return r, c.Complete(r)
+	cc, err := c.Build(r)
+	if err != nil {
+		return nil, err
+	}
+
+	// internal action
+	r.AddAction(
+		newDynamicWatchAction(
+			func(obj client.Object, eventHandler handler.EventHandler, predicates ...predicate.Predicate) error {
+				return cc.Watch(source.Kind(b.mgr.GetCache(), obj), eventHandler, predicates...)
+			},
+			b.watches,
+		),
+	)
+
+	return r, nil
 }
 
 func (b *ReconcilerBuilder) BuildService(ctx context.Context) (*ServiceReconciler, error) {
@@ -246,7 +302,7 @@ func (b *ReconcilerBuilder) BuildService(ctx context.Context) (*ServiceReconcile
 	if !ok {
 		return nil, errors.New("invalid type for object")
 	}
-	r, err := NewServiceReconciler(ctx, b.mgr, name, obj)
+	r, err := NewServiceReconciler(b.mgr, name, obj)
 	if err != nil {
 		return nil, fmt.Errorf("failed to create reconciler for component %s: %w", name, err)
 	}
@@ -267,34 +323,28 @@ func (b *ReconcilerBuilder) BuildService(ctx context.Context) (*ServiceReconcile
 	c = c.For(b.input.object, forOpts...)
 
 	for i := range b.watches {
-		watchOpts := b.watches[i].options
-		if len(watchOpts) == 0 {
-			watchOpts = append(watchOpts, builder.WithPredicates(predicate.And(
-				generation.New(),
-				component.ForLabel(labels.ServicePartOf, b.ownerName),
-			)))
+		if b.watches[i].owned {
+			kinds, _, err := b.mgr.GetScheme().ObjectKinds(b.watches[i].object)
+			if err != nil {
+				return nil, err
+			}
+
+			for i := range kinds {
+				r.AddOwnedType(kinds[i])
+			}
 		}
 
-		c = c.Watches(b.watches[i].object, b.watches[i].eventHandler, watchOpts...)
-	}
-
-	for i := range b.owns {
-		ownOpts := b.owns[i].options
-		if len(ownOpts) == 0 {
-			ownOpts = append(ownOpts, builder.WithPredicates(predicate.And(
-				generation.New(),
-			)))
+		// if the watch is dynamic, then the watcher will be registered
+		// at later stage
+		if b.watches[i].dynamic {
+			continue
 		}
 
-		c = c.Owns(b.owns[i].object, ownOpts...)
-		kinds, _, err := b.mgr.GetScheme().ObjectKinds(b.owns[i].object)
-		if err != nil {
-			return nil, err
-		}
-
-		for i := range kinds {
-			r.AddOwnedType(kinds[i])
-		}
+		c = c.Watches(
+			b.watches[i].object,
+			b.watches[i].eventHandler,
+			builder.WithPredicates(b.watches[i].predicates...),
+		)
 	}
 
 	for i := range b.predicates {
@@ -308,5 +358,13 @@ func (b *ReconcilerBuilder) BuildService(ctx context.Context) (*ServiceReconcile
 		r.AddFinalizer(b.finalizers[i])
 	}
 
-	return r, c.Complete(r)
+	cc, err := c.Build(r)
+	if err != nil {
+		return nil, err
+	}
+
+	// internal action
+	r.AddAction(newDynamicWatchAction(b.mgr, cc, b.watches))
+
+	return r, nil
 }
